requirements.txt
flask
torch
torchvision
numpy
opencv-python
pillow
pycryptodome
scikit-image
matplotlib